requirements.txt
pandas
openpyxl
rapidfuzz
streamlit
google-generativeai
python-dotenv
